views and serializers renamed
