+++ PerformanceTab/view_stats.py
